responsiveness improved and moved button behind the navbar
