+++ src/components/sidebar-extender.js
@@ -1,6 +1,7 @@
 import React, { useState } from "react"
 import "../styles/sidebar-extender.css"
 
+window.sidebarOpen= true;
 
 const SidebarExtender = () => {
   if (typeof window !== 'undefined') {
@@ -21,9 +22,22 @@ const SidebarExtender = () => {
       document.getElementById("desktop-hamburger-icon" ).style.marginLeft= "-100vw"; 
       document.getElementById("documents-window").style.paddingLeft = "2em"; 
     } else {
-      document.getElementById("sidebar").style.marginLeft = "0";
-      document.getElementById("desktop-hamburger-icon" ).style.marginLeft= "0"; 
-      document.getElementById("documents-window").style.paddingLeft = "23em";    
+      if(window.sidebarOpen) {
+        document.getElementById("sidebar").style.marginLeft = "0";
+        document.getElementById("desktop-hamburger-icon" ).style.marginLeft= "0"; 
+        document.getElementById("documents-window").style.paddingLeft = "23em";  
+        document.getElementById("appsody-sidebar-header").style.writingMode = "horizontal-tb"; 
+        document.getElementById("appsody-sidebar-header").style.marginLeft = "1.5rem"; 
+        document.getElementById("docs-sidebar-header").style.writingMode = "horizontal-tb";
+        document.getElementById("docs-sidebar-header").style.marginLeft = "0.25em";
+      } else {
+        document.getElementById("sidebar").style.marginLeft= "-17.6em";
+        document.getElementById("desktopHamburgerOpenbtnId").style.display = "inline"; 
+        document.getElementById("appsody-sidebar-header").style.writingMode = "vertical-rl";
+        document.getElementById("appsody-sidebar-header").style.marginLeft = "9.3em"; 
+        document.getElementById("docs-sidebar-header").style.writingMode = "vertical-rl";
+        document.getElementById("docs-sidebar-header").style.marginLeft = "9.3em"; 
+      }
     }
   }
   const [isExpanded, setIsExpanded] = useState(false);
@@ -32,6 +46,7 @@ const SidebarExtender = () => {
     document.getElementById("hamburger-icon").classList.toggle('open');
 
     setIsExpanded(!isExpanded) 
+    window.sidebarOpen = !isExpanded;
     if (!isExpanded) {
       document.getElementById("sidebar").style.marginLeft= "0";      
     } else {
