+++ src/styles/desktopSidebarCloseButton.css
